Moves LineFragment definition to a separated and dedicated types file
