+++ react/typings/placeComponentsTypes.d.ts
@@ -1,0 +1,7 @@
+import { Address } from 'vtex.checkout-graphql'
+
+export interface LineFragment {
+  name: keyof Address
+  delimiter?: string
+  delimiterAfter?: string
+}
